move figure into wrappers file
